feat: add sortBy in fetchPost
