+++ src/feature/post/ui/PostTable.tsx
@@ -11,6 +11,8 @@ import {
 } from "../../../shared/ui"
 import { usePostForm } from "../model/hook"
 import { useSelectedPostStore } from "../model/store"
+import { Post } from "../type"
+import { useComment } from "../../comment/model/hook"
 
 export const PostTable = ({
   searchQuery,
@@ -18,17 +20,23 @@ export const PostTable = ({
   setSelectedTag,
   updateURL,
   openUserModal,
-  openPostDetail,
 }: {
   searchQuery: string
   selectedTag: string
   setSelectedTag: React.Dispatch<React.SetStateAction<string>>
   updateURL: () => void
   openUserModal: (user: any) => Promise<void>
-  openPostDetail: (post: any) => void
 }) => {
-  const { setSelectedPost, setShowEditDialog, posts } = useSelectedPostStore()
+  const { fetchComments } = useComment()
+  const { setSelectedPost, setShowEditDialog, posts, setShowPostDetailDialog } = useSelectedPostStore()
   const { deletePost } = usePostForm()
+
+  // 게시물 상세 보기
+  const openPostDetail = (post: Post) => {
+    setSelectedPost(post)
+    fetchComments(post.id)
+    setShowPostDetailDialog(true)
+  }
 
   return (
     <Table>
